+++ src/main/kotlin/io/github/jamalam/extensions/user/QuoteExtension.kt
@@ -29,6 +29,7 @@ import com.kotlindiscord.kord.extensions.types.respond
 import dev.kord.common.annotation.KordPreview
 import dev.kord.common.entity.ChannelType
 import dev.kord.common.entity.Snowflake
+import dev.kord.core.behavior.channel.ChannelBehavior
 import dev.kord.core.behavior.channel.createEmbed
 import dev.kord.core.entity.Guild
 import dev.kord.core.entity.Message
@@ -84,7 +85,8 @@ class QuoteExtension : Extension() {
                             arguments.quote,
                             arguments.author.username,
                             arguments.author.avatar!!.url,
-                            user.asUser()
+                            user.asUser(),
+                            channel.asChannel()
                         )
 
                         respond {
@@ -118,7 +120,14 @@ class QuoteExtension : Extension() {
 
                 action {
                     if (guild!!.getConfig().quotesConfig.channel != null) {
-                        sendQuote(this.guild!!.asGuild(), arguments.quote, arguments.author, null, user.asUser())
+                        sendQuote(
+                            this.guild!!.asGuild(),
+                            arguments.quote,
+                            arguments.author,
+                            null,
+                            user.asUser(),
+                            channel.asChannel()
+                        )
                         respond {
                             embed {
                                 info("Quote recorded")
@@ -211,7 +220,14 @@ class QuoteExtension : Extension() {
         }
     }
 
-    private suspend fun sendQuote(guild: Guild, quote: String, quoteAuthor: String, authorIcon: String?, quoter: User) {
+    private suspend fun sendQuote(
+        guild: Guild,
+        quote: String,
+        quoteAuthor: String,
+        authorIcon: String?,
+        quoter: User,
+        commandChannel: ChannelBehavior
+    ) {
         val conf = database.config.getConfig(guild.id)
 
         if (conf.quotesConfig.channel != null) {
@@ -237,6 +253,7 @@ class QuoteExtension : Extension() {
                     userAuthor(quoter.asUser())
                     now()
                     log()
+                    channelField("Channel", commandChannel.asChannel())
                     stringField("Quote", quote)
                     stringField("Author", quoteAuthor)
                 }
@@ -273,6 +290,7 @@ class QuoteExtension : Extension() {
                     userAuthor(quoter.asUser())
                     now()
                     log()
+                    channelField("Channel", this@quote.channel.asChannel())
                     stringField("Quote", content)
                     userField("Author", author2ElectricBoogaloo)
                 }