+++ CHANGELOG.md
@@ -7,10 +7,22 @@
 ### Fixes
 
 - Bots being unmuted will now be logged.
+- Message edited, quote sent and `/ask` used logs now include the relevant channel.
 
 ### Features
 
 - You can no longer quote yourself/your own messages.
+- New logging features:
+  - Reaction removed.
+  - Message scheduled.
+  - Member nickname updated.
+  - Member avatar updated.
+  - Member roles added/removed.
+  - Guild role added/updated/removed.
+  - Message pinned/unpinned.
+  - Message published.
+- Durations in logs are now more readable.
+- KordEx built-in error responses now use embeds.
 
 ## Version 0.7.0